fix: Allow item highlights to work for ConditionalStep
This should ensure that a parent step with requirements like DetailedOwnerStep and ConditionalStep have access to item highlighting. Also simplifies the divide between teleport item highlighting and normal item highlighting.
