+++ src/main/java/com/questhelper/steps/QuestStep.java
@@ -27,6 +27,8 @@ package com.questhelper.steps;
 import com.google.inject.Binder;
 import com.google.inject.Inject;
 import com.google.inject.Module;
+import com.questhelper.requirements.item.ItemRequirement;
+import com.questhelper.steps.tools.QuestPerspective;
 import com.questhelper.steps.widget.AbstractWidgetHighlight;
 import com.questhelper.steps.widget.Spell;
 import com.questhelper.steps.widget.SpellWidgetHighlight;
@@ -45,9 +47,11 @@ import com.questhelper.steps.choice.WidgetTextChange;
 import com.questhelper.steps.choice.WidgetChoiceStep;
 import com.questhelper.steps.choice.WidgetChoiceSteps;
 import com.questhelper.steps.overlay.IconOverlay;
-import java.awt.Graphics2D;
+
+import java.awt.*;
 import java.awt.image.BufferedImage;
 import java.util.*;
+import java.util.List;
 import java.util.regex.Pattern;
 
 import lombok.Getter;
@@ -57,10 +61,13 @@ import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.ChatMessageType;
 import net.runelite.api.Client;
 import net.runelite.api.SpriteID;
+import net.runelite.api.coords.WorldPoint;
 import net.runelite.api.events.ChatMessage;
 import net.runelite.api.events.VarbitChanged;
 import net.runelite.api.events.WidgetLoaded;
+import net.runelite.api.widgets.ComponentID;
 import net.runelite.api.widgets.InterfaceID;
+import net.runelite.api.widgets.Widget;
 import net.runelite.client.callback.ClientThread;
 import net.runelite.client.eventbus.Subscribe;
 import net.runelite.client.game.ItemManager;
@@ -69,6 +76,8 @@ import net.runelite.client.ui.overlay.components.LineComponent;
 import net.runelite.client.ui.overlay.components.PanelComponent;
 import net.runelite.client.ui.overlay.outline.ModelOutlineRenderer;
 import net.runelite.client.ui.overlay.tooltip.TooltipManager;
+import net.runelite.client.util.ColorUtil;
+import net.runelite.client.util.ImageUtil;
 
 @Slf4j
 public abstract class QuestStep implements Module
@@ -571,6 +580,79 @@ public abstract class QuestStep implements Module
 		{
 			renderHoveredMenuEntryPanel(panelComponent, tooltipText);
 		}
+	}
+
+	protected Widget getInventoryWidget()
+	{
+		return client.getWidget(ComponentID.INVENTORY_CONTAINER);
+	}
+
+	protected void renderInventory(Graphics2D graphics, WorldPoint worldPoint, List<ItemRequirement> passedRequirements, boolean distanceLimit)
+	{
+		Widget inventoryWidget = getInventoryWidget();
+		if (inventoryWidget == null || inventoryWidget.isHidden())
+		{
+			return;
+		}
+
+		Color baseColor = questHelper.getConfig().targetOverlayColor();
+
+		if (passedRequirements.isEmpty()) return;
+		if (inventoryWidget.getDynamicChildren() == null) return;
+
+
+		for (Widget item : inventoryWidget.getDynamicChildren())
+		{
+			for (Requirement requirement : passedRequirements)
+			{
+				if (distanceLimit)
+				{
+					WorldPoint playerLocation = client.getLocalPlayer().getWorldLocation();
+					WorldPoint goalWp = QuestPerspective.getInstanceWorldPointFromReal(client, worldPoint);
+					if (goalWp != null && playerLocation.distanceTo(goalWp) <= 100) continue;
+				}
+
+				if (isValidRequirementForRenderInInventory(requirement, item))
+				{
+					highlightInventoryItem(item, baseColor, graphics);
+				}
+			}
+		}
+	}
+
+	private void highlightInventoryItem(Widget item, Color color, Graphics2D graphics)
+	{
+		Rectangle slotBounds = item.getBounds();
+		switch (questHelper.getConfig().highlightStyleInventoryItems())
+		{
+			case SQUARE:
+				graphics.setColor(ColorUtil.colorWithAlpha(color, 65));
+				graphics.fill(slotBounds);
+				graphics.setColor(color);
+				graphics.draw(slotBounds);
+				break;
+			case OUTLINE:
+				BufferedImage outlined = itemManager.getItemOutline(item.getItemId(), item.getItemQuantity(), color);
+				graphics.drawImage(outlined, (int) slotBounds.getX(), (int) slotBounds.getY(), null);
+				break;
+			case FILLED_OUTLINE:
+				BufferedImage outline = itemManager.getItemOutline(item.getItemId(), item.getItemQuantity(), color);
+				graphics.drawImage(outline, (int) slotBounds.getX(), (int) slotBounds.getY(), null);
+				Image image = ImageUtil.fillImage(itemManager.getImage(item.getItemId(), item.getItemQuantity(), false), ColorUtil.colorWithAlpha(color, 65));
+				graphics.drawImage(image, (int) slotBounds.getX(), (int) slotBounds.getY(), null);
+				break;
+			default:
+		}
+	}
+
+	private boolean isValidRequirementForRenderInInventory(Requirement requirement, Widget item)
+	{
+		return requirement instanceof ItemRequirement && isValidRenderRequirementInInventory((ItemRequirement) requirement, item);
+	}
+
+	private boolean isValidRenderRequirementInInventory(ItemRequirement requirement, Widget item)
+	{
+		return requirement.shouldHighlightInInventory(client) && requirement.getAllIds().contains(item.getItemId());
 	}
 
 	protected void renderHoveredItemTooltip(String tooltipText)
