+++ src/main/java/com/questhelper/steps/tools/QuestPerspective.java
@@ -122,6 +122,7 @@ public class QuestPerspective
 
 	public static WorldPoint getInstanceWorldPointFromReal(Client client, WorldPoint wp)
 	{
+		if (wp == null) return null;
 		Collection<WorldPoint> points = QuestPerspective.toLocalInstanceFromReal(client, wp);
 
 		for (WorldPoint point : points)
